refactor: change all faq related files and paths naming conventions
