+++ src/components/sections/FAQ/FAQ.tsx
